+++ SOC_Particle/src/Sync.cpp
@@ -31,14 +31,14 @@ extern SavedPars sp;  // Various parameters to be static at system level and sav
 Sync::Sync()
   : delay_(0), last_(0ULL), now_(0ULL), stat_(false), updateDiff_(0), updateTime_(0)
 {}
-Sync::Sync(unsigned long long delay)
+Sync::Sync(uint64_t delay)
     : delay_(delay), last_(0ULL), now_(0ULL), stat_(false), updateDiff_(0), updateTime_(0)
 {
   updateTimeInput_ = float(delay_)/1000.f;
 }
 
 // Check and count 
-bool Sync::update(bool reset, unsigned long long now, bool andCheck)
+bool Sync::update(bool reset, uint64_t now, bool andCheck)
 {
   now_ = now;
   updateDiff_ = now_ - last_;
@@ -50,7 +50,7 @@ bool Sync::update(bool reset, unsigned long long now, bool andCheck)
   }
   return( stat_ );
 }
-bool Sync::update(unsigned long long now, bool reset, bool andCheck)
+bool Sync::update(uint64_t now, bool reset, bool andCheck)
 {
   now_ = now;
   updateDiff_ = now_ - last_;
@@ -62,7 +62,7 @@ bool Sync::update(unsigned long long now, bool reset, bool andCheck)
   }
   return( stat_ );
 }
-bool Sync::update(unsigned long long now, bool reset)
+bool Sync::update(uint64_t now, bool reset)
 {
   now_ = now;
   updateDiff_ = now_ - last_;
@@ -74,7 +74,7 @@ bool Sync::update(unsigned long long now, bool reset)
   }
   return( stat_ );
 }
-bool Sync::updateN(unsigned long long now, bool reset, bool orCheck)
+bool Sync::updateN(uint64_t now, bool reset, bool orCheck)
 {
   now_ = now;
   updateDiff_ = now_ - last_;
