+++ SOC_Particle/src/serial.h
@@ -43,7 +43,7 @@ extern CommandPars cp;  // Various parameters to be static at system level
 
 // Headers
 String chat_cmd_from(String *source);
-void delay_no_block(const unsigned long long int delay_millis);
+void delay_no_block(const uint64_t delay_millis);
 String finish_request(const String in_str);
 bool is_finished(const char in_char);
 void print_battery_header();
